+++ src/DotPulsar/Internal/Connection.cs
@@ -21,6 +21,7 @@ using DotPulsar.Internal.Abstractions;
 using DotPulsar.Internal.Exceptions;
 using DotPulsar.Internal.Extensions;
 using DotPulsar.Internal.PulsarApi;
+using Google.Protobuf;
 using System.Buffers;
 
 public sealed class Connection : IConnection
@@ -101,7 +102,7 @@ public sealed class Connection : IConnection
         {
             command.Response ??= new AuthData();
             command.Response.AuthMethodName = _authentication.AuthenticationMethodName;
-            command.Response.Data = await _authentication.GetAuthenticationData(cancellationToken).ConfigureAwait(false);
+            command.Response.AuthData_ = ByteString.CopyFrom(await _authentication.GetAuthenticationData(cancellationToken).ConfigureAwait(false));
         }
 
         await Send(command.AsBaseCommand(), cancellationToken).ConfigureAwait(false);
@@ -145,7 +146,7 @@ public sealed class Connection : IConnection
         if (_authentication is not null)
         {
             command.AuthMethodName = _authentication.AuthenticationMethodName;
-            command.AuthData = await _authentication.GetAuthenticationData(cancellationToken).ConfigureAwait(false);
+            command.AuthData = ByteString.CopyFrom(await _authentication.GetAuthenticationData(cancellationToken).ConfigureAwait(false));
         }
 
         Task<BaseCommand>? responseTask;
@@ -259,7 +260,7 @@ public sealed class Connection : IConnection
         }
         catch (OperationCanceledException)
         {
-            _ = responseTcs.TrySetCanceled();
+            _ = responseTcs.TrySetCanceled(CancellationToken.None);
             throw;
         }
     }
@@ -357,17 +358,17 @@ public sealed class Connection : IConnection
             await foreach (var frame in _stream.Frames(cancellationToken).ConfigureAwait(false))
             {
                 var commandSize = frame.ReadUInt32(0, true);
-                var command = Serializer.Deserialize<BaseCommand>(frame.Slice(4, commandSize));
+                var command = BaseCommand.Parser.ParseFrom(frame.Slice(4, commandSize));
 
-                _pingPongHandler.Incoming(command.CommandType);
+                _pingPongHandler.Incoming(command.Type);
 
-                if (command.CommandType == BaseCommand.Type.Message)
+                if (command.Type == BaseCommand.Types.Type.Message)
                     _channelManager.Incoming(command.Message, new ReadOnlySequence<byte>(frame.Slice(commandSize + 4).ToArray()));
-                else if (command.CommandType == BaseCommand.Type.AuthChallenge)
+                else if (command.Type == BaseCommand.Types.Type.AuthChallenge)
                     _ = Task.Factory.StartNew(async () => await Send(new CommandAuthResponse(), cancellationToken).ConfigureAwait(false));
-                else if (command.CommandType == BaseCommand.Type.Ping)
+                else if (command.Type == BaseCommand.Types.Type.Ping)
                     _ = Task.Factory.StartNew(async () => await Send(new CommandPong(), cancellationToken).ConfigureAwait(false));
-                else if (command.CommandType == BaseCommand.Type.Pong)
+                else if (command.Type == BaseCommand.Types.Type.Pong)
                     continue;
                 else
                     _channelManager.Incoming(command);