+++ src/DotPulsar/Internal/Extensions/MessagePackageExtensions.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -23,7 +23,7 @@ public static class MessagePackageExtensions
         => package.Data.ReadUInt32(Constants.MetadataSizeOffset, true);
 
     public static MessageMetadata ExtractMetadata(this MessagePackage package, uint metadataSize)
-        => Serializer.Deserialize<MessageMetadata>(package.Data.Slice(Constants.MetadataOffset, metadataSize));
+        => MessageMetadata.Parser.ParseFrom(package.Data.Slice(Constants.MetadataOffset, metadataSize));
 
     public static ReadOnlySequence<byte> ExtractData(this MessagePackage package, uint metadataSize)
         => package.Data.Slice(Constants.MetadataOffset + metadataSize);